+++ netring-flow-dns/src/lib.rs
@@ -37,11 +37,13 @@
 #![cfg_attr(docsrs, feature(doc_cfg))]
 
 mod correlator;
+mod datagram;
 mod observer;
 mod parser;
 mod types;
 
 pub use correlator::Correlator;
+pub use datagram::{DnsMessage, DnsUdpParser};
 pub use observer::DnsUdpObserver;
 pub use parser::{DnsParseResult, parse_message, parse_message_at};
 pub use types::*;