+++ netring-flow-http/src/lib.rs
@@ -40,8 +40,10 @@
 
 mod factory;
 mod parser;
+mod session;
 mod types;
 
 pub use factory::{HttpFactory, HttpReassembler};
 pub use parser::Error;
+pub use session::{HttpMessage, HttpParser};
 pub use types::{HttpConfig, HttpHandler, HttpRequest, HttpResponse, HttpVersion};